Update SnackBar ANd Alert Dialog

+++ app/src/main/java/com/skripsi/perpustakaanapp/ui/book/detailbook/DetailBookActivity.kt
@@ -17,11 +17,12 @@ import com.skripsi.perpustakaanapp.core.SessionManager
 import com.skripsi.perpustakaanapp.core.apihelper.RetrofitClient
 import com.skripsi.perpustakaanapp.core.models.Book
 import com.skripsi.perpustakaanapp.core.repository.LibraryRepository
-import com.skripsi.perpustakaanapp.core.resource.Resource
-import com.skripsi.perpustakaanapp.utils.NetworkInfo.BOOK_IMAGE_BASE_URL
+import com.skripsi.perpustakaanapp.core.resource.MyResource
 import com.skripsi.perpustakaanapp.databinding.ActivityDetailBookBinding
 import com.skripsi.perpustakaanapp.ui.MyAlertDialog
+import com.skripsi.perpustakaanapp.ui.MySnackBar
 import com.skripsi.perpustakaanapp.ui.admin.bookmanagerial.updatebook.UpdateBookFragment
+import com.skripsi.perpustakaanapp.utils.NetworkInfo.BOOK_IMAGE_BASE_URL
 import com.skripsi.perpustakaanapp.utils.setSingleClickListener
 
 
@@ -90,7 +91,17 @@ class DetailBookActivity : AppCompatActivity() {
                 true
             }
             R.id.delete_menu -> {
-                deleteBook()
+                MyAlertDialog.showWith2Event(
+                    this,
+                    null,
+                    resources.getString(R.string.delete_confirmation),
+                    resources.getString(R.string.confirmation_yes),
+                    resources.getString(R.string.confirmation_no),
+                    {_,_ ->
+                        deleteBook()
+                    }, {_,_ ->
+
+                    })
                 true
             }
             android.R.id.home -> {
@@ -102,7 +113,11 @@ class DetailBookActivity : AppCompatActivity() {
     }
 
     private fun updateBook() {
+        val bundle = Bundle()
+        bundle.putParcelable(UpdateBookFragment.FRAGMENT_EXTRA_DATA, detailBook)
+
         val bottomDialogFragment = UpdateBookFragment()
+        bottomDialogFragment.arguments = bundle
         bottomDialogFragment.show(supportFragmentManager, "UpdateBookFragment")
     }
 
@@ -114,26 +129,18 @@ class DetailBookActivity : AppCompatActivity() {
         viewModel.resourceDeleteBook.observe(this) { event ->
             event.getContentIfNotHandled()?.let { resource ->
                 when (resource) {
-                    is Resource.Loading -> {
+                    is MyResource.Loading -> {
                         binding.progressBar.visibility = View.VISIBLE
                     }
-                    is Resource.Success -> {
+                    is MyResource.Success -> {
                         binding.progressBar.visibility = View.GONE
-                        MyAlertDialog.showAlertDialogEvent(this,
-                            R.drawable.icon_checked,
-                            resource.data.toString().uppercase(),
-                            "Buku Berhasil Di Hapus")
-                        { _, _ ->
-                            setResult(RESULT_OK) //set return data is "RESULT_OK" after success deleted
-                            finish()
-                        }
+                        MySnackBar.showBlack(binding.root, resource.data.toString())
+                        setResult(RESULT_OK) //set return data is "RESULT_OK" after success deleted
+                        finish()
                     }
-                    is Resource.Error -> {
+                    is MyResource.Error -> {
                         binding.progressBar.visibility = View.GONE
-                        MyAlertDialog.showAlertDialog(this,
-                            R.drawable.icon_cancel,
-                            "FAILED",
-                            resource.message.toString())
+                        MySnackBar.showRed(binding.root, resource.message.toString())
                     }
                 }
             }
@@ -146,17 +153,17 @@ class DetailBookActivity : AppCompatActivity() {
         viewModel.resourceDetailBook.observe(this) { event ->
             event.getContentIfNotHandled()?.let { resource ->
                 when(resource) {
-                    is Resource.Loading -> {
+                    is MyResource.Loading -> {
                         binding.progressBar.visibility = View.VISIBLE
                     }
-                    is Resource.Success -> {
+                    is MyResource.Success -> {
                         detailBook = resource.data
                         showDetailBook()
                         binding.progressBar.visibility = View.GONE
                     }
-                    is Resource.Error -> {
+                    is MyResource.Error -> {
                         binding.progressBar.visibility = View.GONE
-                        MyAlertDialog.showAlertDialog(this, R.drawable.icon_cancel, "Failed", resource.message.toString())
+                        MyAlertDialog.show(this, R.drawable.icon_cancel, "Failed", resource.message.toString())
                     }
                 }
             }
@@ -164,7 +171,7 @@ class DetailBookActivity : AppCompatActivity() {
     }
 
     private fun showDetailBook() {
-        setEnableButton()
+        setEnableLoanButton()
 
         binding.progressBar.visibility = View.GONE
         supportActionBar?.title = detailBook?.title
@@ -200,11 +207,15 @@ class DetailBookActivity : AppCompatActivity() {
         }
     }
 
-    private fun setEnableButton() {
-        if (detailBook?.stock != "0") {
-            doLoan()
+    private fun setEnableLoanButton() {
+        if (sessionManager.fetchUserRole() == "student") {
+            if (detailBook?.stock != "0") {
+                doLoan()
+            } else {
+                binding.buttonLoan.isEnabled = false
+            }
         } else {
-            binding.buttonLoan.isEnabled = false
+            binding.buttonLoan.visibility = View.INVISIBLE
         }
     }
 
@@ -215,13 +226,13 @@ class DetailBookActivity : AppCompatActivity() {
             viewModel.resourceLoanBook.observe(this) { event ->
                 event.getContentIfNotHandled()?.let {
                     when (it) {
-                        is Resource.Loading -> {
+                        is MyResource.Loading -> {
                             binding.progressBar.visibility = View.VISIBLE
                         }
 
-                        is Resource.Success -> {
+                        is MyResource.Success -> {
                             binding.progressBar.visibility = View.GONE
-                            MyAlertDialog.showAlertDialogEvent(this@DetailBookActivity,
+                            MyAlertDialog.showWithEvent(this@DetailBookActivity,
                                 R.drawable.icon_checked,
                                 it.data.toString().uppercase(),
                                 "Menunggu Admin")
@@ -230,9 +241,9 @@ class DetailBookActivity : AppCompatActivity() {
                             }
                         }
 
-                        is Resource.Error -> {
+                        is MyResource.Error -> {
                             binding.progressBar.visibility = View.GONE
-                            MyAlertDialog.showAlertDialog(this@DetailBookActivity,
+                            MyAlertDialog.show(this@DetailBookActivity,
                                 R.drawable.icon_cancel,
                                 "FAILED",
                                 it.message.toString())
@@ -243,13 +254,15 @@ class DetailBookActivity : AppCompatActivity() {
         }
     }
 
-    private  fun observeStatusFavorite(bookId: String?) {
-        viewModel.statusFavorite(sessionManager.fetchAuthToken().toString(), sessionManager.fetchUsername().toString(), bookId)
+    private fun observeStatusFavorite(bookId: String?) {
+        if (sessionManager.fetchUserRole() == "student") {
+            viewModel.statusFavorite(sessionManager.fetchAuthToken().toString(), sessionManager.fetchUsername().toString(), bookId)
 
-        viewModel.isFavorite.observe(this) { event ->
-            event.getContentIfNotHandled().let {
-                setStatusFavorite(it)
+            viewModel.isFavorite.observe(this) { event ->
+                setStatusFavorite(event.getContentIfNotHandled())
             }
+        } else {
+            binding.fabFavorite.visibility = View.INVISIBLE
         }
     }
 
@@ -273,28 +286,28 @@ class DetailBookActivity : AppCompatActivity() {
     }
 
     private fun favoriteViewModel(isFavorite: Boolean) {
-        val bookId: String?
-        if (detailBook?.bookId == null) {
-            bookId = intent.getStringExtra(BOOK_ID)
-        }
-        else {
-            bookId = detailBook?.bookId
-        }
+        val bookId: String? =
+            if (detailBook?.bookId == null) {
+                intent.getStringExtra(BOOK_ID)
+            } else {
+                detailBook?.bookId
+            }
+
         viewModel.changeStatusFavorite(sessionManager.fetchAuthToken().toString(), sessionManager.fetchUsername(), bookId, isFavorite)
 
         viewModel.resourceChangeStatusFavorite.observe(this) { event ->
             event.getContentIfNotHandled().let { resource ->
                 when(resource) {
-                    is Resource.Loading -> {
+                    is MyResource.Loading -> {
                         binding.progressBar.visibility = View.VISIBLE
                     }
-                    is Resource.Success -> {
+                    is MyResource.Success -> {
                         binding.progressBar.visibility = View.GONE
                         Snackbar
                             .make(binding.root, resource.data.toString(), Snackbar.LENGTH_LONG)
                             .show()
                     }
-                    is Resource.Error -> {
+                    is MyResource.Error -> {
                         binding.progressBar.visibility = View.GONE
                         Snackbar
                             .make(binding.root, resource.message.toString(), Snackbar.LENGTH_LONG)
@@ -310,215 +323,3 @@ class DetailBookActivity : AppCompatActivity() {
         const val BOOK_ID = "book_id"
     }
 }
-
-
-
-
-/*
-package com.skripsi.perpustakaanapp.ui.user
-
-import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
-import android.view.View
-import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.ViewModelProvider
-import com.skripsi.perpustakaanapp.R
-import com.skripsi.perpustakaanapp.core.MViewModelFactory
-import com.skripsi.perpustakaanapp.core.SessionManager
-import com.skripsi.perpustakaanapp.core.apihelper.RetrofitClient
-import com.skripsi.perpustakaanapp.core.models.Book
-import com.skripsi.perpustakaanapp.core.repository.LibraryRepository
-import com.skripsi.perpustakaanapp.core.resource.Resource
-import com.skripsi.perpustakaanapp.databinding.ActivityDetailBookBinding
-import com.skripsi.perpustakaanapp.ui.MyAlertDialog
-import com.skripsi.perpustakaanapp.ui.admin.book.UpdateBookActivity
-
-class DetailBookActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityDetailBookBinding
-    private lateinit var sessionManager: SessionManager
-    private lateinit var viewModel: DetailBookViewModel
-
-    private var detailBook: Book? = null
-    private val client = RetrofitClient
-
-    var activityDetailBook: DetailBookActivity? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityDetailBookBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        supportActionBar?.setDisplayShowHomeEnabled(true)
-
-        sessionManager = SessionManager(this)
-
-        activityDetailBook = this@DetailBookActivity
-
-        viewModel= ViewModelProvider(this,MViewModelFactory(LibraryRepository(client))).get(
-            DetailBookViewModel::class.java
-        )
-
-        detailBook = intent.getParcelableExtra<Book>(EXTRA_DATA)
-        showDetailBook()
-
-        //if buku belum ada yang pinjam
-        //binding.buttonLoan.isEnabled = true
-        //else
-        //binding.buttonLoan.isEnabled = false
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        if (sessionManager.fetchUserRole() == "admin") {
-            menuInflater.inflate(R.menu.activity_detail_menu, menu)
-        }
-        return super.onCreateOptionsMenu(menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return when (item.itemId) {
-            R.id.edit_menu -> {
-                updateBook()
-                true
-            }
-            R.id.delete_menu -> {
-                deleteBook()
-                true
-            }
-            else -> true
-        }
-    }
-
-    private fun updateBook() {
-        val b = UpdateBookActivity()
-        b.show(supportFragmentManager, "Hi")
-    }
-
-    private fun deleteBook(){
-        detailBook?.bookId?.let {
-            viewModel.deleteBook(token = "Bearer ${sessionManager.fetchAuthToken()}", it)
-        }
-
-//        viewModel.responseMessage.observe(this) { message ->
-//            if (message != null) {
-////                var i = 0
-////                while (i<=10) {
-////                    println("nilai message = $message")
-////                    i++
-////                }
-//                //Reset status value at first to prevent multitriggering
-//                //and to be available to trigger action again
-//                viewModel.responseMessage.value = null
-//                if (message == "success") {
-//                    MyAlertDialog.showAlertDialogEvent(this@DetailBookActivity, R.drawable.icon_checked, "SUCCESS", "Buku Berhasil Di Hapus") { _, _ ->
-//                        finish()
-//                    }
-//                }
-//                else {
-//                    MyAlertDialog.showAlertDialog(this@DetailBookActivity, R.drawable.icon_cancel, "FAILED", message)
-//                }
-//            }
-//        }
-
-        viewModel.deleteBook.observe(this) { event ->
-            event.getContentIfNotHandled()?.let {
-               when (it) {
-                   is  Resource.Loading -> {
-                       binding.progressBar.visibility = View.VISIBLE
-                   }
-                   is Resource.Success -> {
-                       binding.progressBar.visibility = View.GONE
-                       MyAlertDialog.showAlertDialogEvent(this@DetailBookActivity, R.drawable.icon_checked, it.data.toString().uppercase(), "Buku Berhasil Di Hapus")
-                       { _, _ ->
-                            finish()
-                       }
-                   }
-                   is Resource.Error -> {
-                       binding.progressBar.visibility = View.GONE
-                       MyAlertDialog.showAlertDialog(this@DetailBookActivity, R.drawable.icon_cancel, "FAILED", it.message.toString())
-                   }
-               }
-            }
-        }
-
-//        viewModel.errorMessage.observe(this) {
-//            if (it != null) {
-//                MyAlertDialog.showAlertDialog(this@DetailBookActivity, R.drawable.icon_cancel, "ERROR", it)
-//                viewModel.errorMessage.value = null
-//            }
-//        }
-    }
-
-    private fun showDetailBook() {
-            setEnableButton()
-            binding.progressBar.visibility = View.GONE
-            supportActionBar?.title = detailBook?.title
-            binding.author.text = detailBook?.author
-            binding.year.text = detailBook?.edition  //harusnya tahun terbit
-            binding.publisher.text = detailBook?.publisher
-
-    }
-
-    private fun setEnableButton() {
-        if (detailBook?.copies != "0") {
-            doLoan()
-        } else {
-            binding.buttonLoan.isEnabled = false
-        }
-    }
-
-    private fun doLoan() {
-        binding.buttonLoan.setOnClickListener {
-            viewModel.createTransaction(sessionManager.fetchUserName(), detailBook?.bookId)
-
-            println("Disini ${sessionManager.fetchUserName()}, ${detailBook?.bookId}")
-
-//            viewModel.responseMessage2.observe(this) { messsage ->
-//                if (messsage != null) {
-//                    //Reset status value at first to prevent multitriggering
-//                    //and to be available to trigger action again
-//                    viewModel.responseMessage2.value = null
-//                    if (messsage == "success") {
-//                        MyAlertDialog.showAlertDialog(this@DetailBookActivity, R.drawable.icon_checked, messsage.uppercase(Locale.getDefault()), "Berhasil Dan Mohon Menunggu Persetujuan Admin")
-//                        binding.buttonLoan.isEnabled = false
-//                    }
-//                    else {
-//                        MyAlertDialog.showAlertDialog(this@DetailBookActivity, R.drawable.icon_cancel, "FAILED", messsage)
-//                    }
-//                }
-//            }
-            viewModel.loanBook.observe(this) { event ->
-                event.getContentIfNotHandled()?.let {
-                    when (it) {
-                        is Resource.Loading -> {
-                            binding.progressBar.visibility = View.VISIBLE
-                        }
-                        is Resource.Success -> {
-                            binding.progressBar.visibility = View.GONE
-                            MyAlertDialog.showAlertDialogEvent(this@DetailBookActivity, R.drawable.icon_checked, it.data.toString().uppercase(), "Menunggu Admin")
-                            { _, _ ->
-                                binding.buttonLoan.isEnabled = false
-                            }
-                        }
-                        is Resource.Error -> {
-                            binding.progressBar.visibility = View.GONE
-                            MyAlertDialog.showAlertDialog(this@DetailBookActivity, R.drawable.icon_cancel, "FAILED", it.message.toString())
-                        }
-                    }
-                }
-            }
-
-//            viewModel.errorMessage.observe(this) {
-//                if (it != null) {
-//                    MyAlertDialog.showAlertDialog(this@DetailBookActivity, R.drawable.icon_cancel, "ERROR", "Mohon Maaf Silahkan Coba Beberapa Saat Lagi")
-//                    viewModel.responseMessage.value = null
-//                }
-//            }
-        }
-    }
-
-    companion object {
-        const val EXTRA_DATA = "extra_data"
-    }
-}
- */
